Make turtle creation functions take Doubles as (x an y) inputs.

+++ src/main/scala/net/kogics/kojo/core/TSCanvasFeatures.scala
@@ -21,7 +21,7 @@ import java.awt.Color
 
 trait TSCanvasFeatures {
   def clear(): Unit
-  def newTurtle(x: Int, y: Int): Turtle
+  def newTurtle(x: Double, y: Double): Turtle
   def axesOn(): Unit
   def axesOff(): Unit
   def gridOn(): Unit
